reorganization completed, all seems to be working
