+++ src/core/utilities.c
@@ -6,8 +6,9 @@
 
 #include <zmq.h>
 
-#include "include/dtc.h"
-#include "utilities.h"
+#include <dtc.h>
+#include "include/logger.h"
+#include "include/utilities.h"
 
 int lookup_uint16_conf_element(const config_setting_t *setting,
                                       const char *name, uint16_t *out)
